chore : reorganization of directory structure
