+++ src/main/java/com/Techeer/Team_C/domain/user/controller/UserController.java
@@ -2,10 +2,10 @@ package com.Techeer.Team_C.domain.user.controller;
 
 import com.Techeer.Team_C.domain.user.dto.SignupFormDto;
 import com.Techeer.Team_C.domain.user.dto.UserDto;
-import com.Techeer.Team_C.domain.user.auth.JwtTokenProvider;
+import com.Techeer.Team_C.domain.auth.jwt.JwtTokenProvider;
 import com.Techeer.Team_C.domain.user.repository.UserRepository;
 import com.Techeer.Team_C.domain.user.entity.User;
-import com.Techeer.Team_C.domain.user.service.AuthService;
+import com.Techeer.Team_C.domain.auth.service.AuthService;
 import com.Techeer.Team_C.global.error.exception.BusinessException;
 import com.Techeer.Team_C.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
