+++ src/main/java/com/Techeer/Team_C/global/config/WebSecurityConfig.java
@@ -1,7 +1,7 @@
 package com.Techeer.Team_C.global.config;
 
-import com.Techeer.Team_C.domain.user.auth.JwtAuthenticationFilter;
-import com.Techeer.Team_C.domain.user.auth.JwtTokenProvider;
+import com.Techeer.Team_C.domain.auth.jwt.JwtAuthenticationFilter;
+import com.Techeer.Team_C.domain.auth.jwt.JwtTokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.authentication.AuthenticationManager;
